+++ CHANGELOG.md
@@ -65,6 +65,7 @@ Available in the `dev` branch
 - `tabview` by default allow auto expanding the page only to right and bottom (#1573)
 - fix crash when drawing gradient to the same color
 - chart: fix memory leak
+- `img`: improve hit test for transformed images
 
 ## v7.0.1 (01.06.2020)